requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
DateTime==5.5
et-xmlfile==1.1.0
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
numpy==1.24.4
openpyxl==3.1.5
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
protobuf==5.27.2
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
streamlit==1.36.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zipp==3.19.2
zope.interface==6.4.post2
